refactor: abstract DomainDetail cmp
