+++ src/components/common/DomainDetail/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './cmp'
