+++ src/components/common/DomainDetail/types.tsx
@@ -1,0 +1,4 @@
+export type DomainDetailProps = {
+  domainId: string
+  showDelete?: boolean
+}
